Modified Connectable::checkPlayer to use CLOCKS_PER_SEC to fix login bug in systems with a CLOCK_PER_SEC value other than 1000. Modified the loading of skills so that level and maxlevel are cast to uint8_t.
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@1642 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/PlayerSkills.cpp
@@ -64,8 +64,8 @@ void PlayerSkills::load() {
 	mysqlpp::StoreQueryResult res = query.store();
 	for (size_t i = 0; i < res.num_rows(); i++) {
 		PlayerSkillInfo skill;
-		skill.level = res[i][1];
-		skill.maxlevel = res[i][2];
+		skill.level = (uint8_t) res[i][1];
+		skill.maxlevel = (uint8_t) res[i][2];
 		playerskills[res[i][0]] = skill;
 	}
 }
